:truck: Reorganize to /shadcn and /sidebar
