+++ apps/next/src/components/ui/sidebar/sidebar-content.tsx
@@ -1,9 +1,9 @@
 import Image from "next/image";
-import { SheetContent, SheetTitle } from "./sheet";
-import { Avatar, AvatarImage, AvatarFallback } from "./avatar";
-import { Button } from "./button";
-import SheetButton from "./sheet-button";
-import SheetDivider from "./sheet-divider";
+import { SheetContent, SheetTitle } from "../shadcn/sheet";
+import { Avatar, AvatarImage, AvatarFallback } from "../shadcn/avatar";
+import { Button } from "../shadcn/button";
+import SidebarButton from "./sidebar-button";
+import SidebarDivider from "./sidebar-divider";
 import { Settings2, CalendarFold, LogOut, House, Info, Pin, Trophy } from "lucide-react";
 
 export default function SidebarContent() {
@@ -24,14 +24,14 @@ export default function SidebarContent() {
                 <span className="text-sm font-normal">v0.1 (preview)</span>
               </SheetTitle>
             </div>
-            <SheetDivider title="Dining Hall Info"/>
-            <SheetButton Icon={House} title="Home" href="/"/>
-            <SheetButton Icon={CalendarFold} title="Events" href="/events"/>
-            <SheetButton Icon={Pin} title="My Favorites" href="/my-favs" deactivated/>
-            <SheetButton Icon={Trophy} title="Most Liked" href="/leaderboard" deactivated/>
-            <SheetDivider title="Miscellaneous"/>
-            <SheetButton Icon={Settings2} title="Settings" href="/settings" deactivated/>
-            <SheetButton Icon={Info} title="About" href="/about"/>
+            <SidebarDivider title="Dining Hall Info"/>
+            <SidebarButton Icon={House} title="Home" href="/"/>
+            <SidebarButton Icon={CalendarFold} title="Events" href="/events"/>
+            <SidebarButton Icon={Pin} title="My Favorites" href="/my-favs" deactivated/>
+            <SidebarButton Icon={Trophy} title="Most Liked" href="/leaderboard" deactivated/>
+            <SidebarDivider title="Miscellaneous"/>
+            <SidebarButton Icon={Settings2} title="Settings" href="/settings" deactivated/>
+            <SidebarButton Icon={Info} title="About" href="/about"/>
           </div>
           <div className="flex p-2 items-center justify-between rounded-md hover:bg-zinc-100 transition-colors" id="sheet-bottom">
             <div className="flex gap-3 items-center">
