+++ apps/next/src/components/ui/sidebar/sidebar-divider.tsx
@@ -1,4 +1,4 @@
-export default function SheetDivider({title} : {title: string}) {
+export default function SidebarDivider({title} : {title: string}) {
     return (
         <span className="text-sm text-zinc-600 mt-4">{title}</span>
     )
